Ability to check entity state, not only whether exists
In some cases, we might need to check not only whether the entity
exists, but also the state (for example is an article published).
During validation, a method to verify the fetched entity is called.
The default implementation always returns true, so when you need custom
implementation you need to overwrite this method.